Change type of ID fields to int64 from int

+++ internal/model/account.go
@@ -7,8 +7,8 @@ import (
 type Account struct {
 	BaseEntity
 	Name          string          `json:"name" validate:"required"`
-	UserId        int             `json:"userId" validate:"required"`
-	ExchangeId    int             `json:"exchangeId" validate:"required"`
+	UserId        int64           `json:"userId" validate:"required"`
+	ExchangeId    int64           `json:"exchangeId" validate:"required"`
 	Capital       decimal.Decimal `json:"capital" validate:"required"`
 	LockedCapital decimal.Decimal `json:"lockedCapital"`
 }
